the worse code ever in every file it exists in
I am so dumb it hurts. Is a catch 22 to call yourself stupid? I think im
stupid. But would a stupid person know their stupid? And if their not
stupid then does calling themself stupid make them stupid because only a
smart person who is stupid would call themself stupid?

The only reason this commit exists is because I feel as If I may give up
on this. As if I cant convert my likely crappy idea concept into
something real. I dont really know what im making, how could i possibly
write tests for this and do things the right way when they take time. Is
the time spent fixing it later greater than doing it right the first
time and every time after the first because its not going to be what I
want the first time since I dont know what it is at this point. Does it
make a difference that I know which parts are bad or at least I think
are bad? Probably not, it is knowledge not applied therefore wasted an
may as well not have been learned in the first place. None of it really
matters though does it since no one is reading this and if they were I
wouldnt be embarassed because im just an idiot that doesnt care anyway.
What does that say about me that I dont care. Maybe I do care but just
dont want to admit it. Why commit this in the first place or even work
on this. Why am I making this. I dont know. Maybe ill figure it out on
the way. The way. Which way is that. What is the way.